requirements.txt
python-dotenv
pandas
numpy
pymongo[srv]>=4.0
scikit-learn
scipy
matplotlib
seaborn
certifi
dill
pyaml
# -e .